Refactor command handling to utilize method references and improve code readability across the game logic.

+++ src/main/java/de/oglimmer/ggo/logic/Field.java
@@ -60,6 +60,6 @@ public class Field implements Serializable {
     }
 
     public String asPosString() {
-        return pos.getX() + ":" + pos.getY();
+        return ((int) pos.getX()) + ":" + ((int) pos.getY());
     }
 }
